Replaced JsonConvert.js "this" referencees with "Newtonsoft.Json.JsonConvert" for consistency and to avoid potential bind confusion
Some places already used "Newtonsoft.Json.JsonConvert" instead of "this"
(eg. calls to "Newtonsoft.Json.JsonConvert.SerializeObject") while some
did use this (eg. calls to "this.stringify") - it seems to me to make
more sense to use one of the other in all cases and removing "this"
means that we never have to worry about what "this" is (which allowed me
to remove the ".bind(this)" call that I added when I created the
getInstanceBuilder function).